# src/GlazeAudio.Api/Endpoints/SongEndpoints.cs
using GlazeAudio.Api.Contracts;
using GlazeAudio.Api.Data;
using GlazeAudio.Api.Infrastructure;
using GlazeAudio.Api.Models;
using Microsoft.EntityFrameworkCore;

namespace GlazeAudio.Api.Endpoints;

public static class SongEndpoints
{
    public static RouteGroupBuilder MapSongEndpoints(this IEndpointRouteBuilder routes)
    {
        var group = routes.MapGroup("/api/albums/{albumId:int}/songs").WithTags("Songs");

        group.MapGet("/", GetAll)
            .WithName("GetSongs")
            .WithSummary("List songs of an album (paged, filterable)")
            .WithDescription(
                "Returns one page of the album's songs ordered by track number. Filter with search, minDuration " +
                "and maxDuration. 404 if the album does not exist.")
            .Produces<PagedResult<SongDto>>()
            .ProducesValidationProblem(StatusCodes.Status400BadRequest)
            .ProducesProblem(StatusCodes.Status404NotFound);

        group.MapGet("/{songId:int}", GetById)
            .WithName("GetSong")
            .WithSummary("Get one song")
            .WithDescription("404 if the album does not exist or the song does not belong to it.")
            .Produces<SongDto>()
            .ProducesProblem(StatusCodes.Status404NotFound);

        group.MapPost("/", Create)
            .WithName("CreateSong")
            .WithSummary("Add a song to an album")
            .WithValidation<SongRequest>()
            .Produces<SongDto>(StatusCodes.Status201Created)
            .ProducesProblem(StatusCodes.Status404NotFound);

        group.MapPut("/{songId:int}", Update)
            .WithName("UpdateSong")
            .WithSummary("Update a song")
            .WithValidation<SongRequest>()
            .Produces<SongDto>()
            .ProducesProblem(StatusCodes.Status404NotFound);

        group.MapDelete("/{songId:int}", Delete)
            .WithName("DeleteSong")
            .WithSummary("Delete a song")
            .WithDescription("Deletes the song and all of its reviews. Returns 204 with no body.")
            .Produces(StatusCodes.Status204NoContent)
            .ProducesProblem(StatusCodes.Status404NotFound);

        return group;
    }

    private static IQueryable<SongDto> Project(IQueryable<Song> songs) =>
        songs.Select(s => new SongDto(
            s.Id, s.AlbumId, s.Title, s.TrackNumber, s.DurationSeconds,
            s.Reviews.Count,
            s.Reviews.Average(r => (double?)(r.LyricsRating + r.MelodyRating + r.MoodRating + r.ExpressivenessRating) / 4.0)));

    private static SongDto WithLinks(SongDto song, HttpRequest request) =>
        song with { Links = ApiLinks.Song(request, song.AlbumId, song.Id) };

    private static async Task<IResult> GetAll(int albumId, [AsParameters] SongQuery query, GlazeAudioDbContext db, HttpRequest request, CancellationToken ct)
    {
        if (Paging.Validate(query.Page, query.PageSize, out var page, out var pageSize) is { } invalid)
            return invalid;
        if (!await db.Albums.AnyAsync(a => a.Id == albumId, ct))
            return ApiProblems.NotFound("Album", albumId);

        var songs = db.Songs.AsNoTracking().Where(s => s.AlbumId == albumId);

        if (!string.IsNullOrWhiteSpace(query.Search))
        {
            var term = query.Search.Trim().ToLower();
            songs = songs.Where(s => s.Title.ToLower().Contains(term));
        }
        if (query.MinDuration is { } minDuration) songs = songs.Where(s => s.DurationSeconds >= minDuration);
        if (query.MaxDuration is { } maxDuration) songs = songs.Where(s => s.DurationSeconds <= maxDuration);

        var total = await songs.CountAsync(ct);
        var items = await Project(songs.OrderBy(s => s.TrackNumber).ThenBy(s => s.Id).Skip((page - 1) * pageSize).Take(pageSize))
            .ToListAsync(ct);

        return Results.Ok(Paging.Create(request, items.Select(s => WithLinks(s, request)).ToList(), page, pageSize, total));
    }

    private static async Task<IResult> GetById(int albumId, int songId, GlazeAudioDbContext db, HttpRequest request, CancellationToken ct)
    {
        if (!await db.Albums.AnyAsync(a => a.Id == albumId, ct))
            return ApiProblems.NotFound("Album", albumId);

        var song = await Project(db.Songs.Where(s => s.AlbumId == albumId && s.Id == songId)).FirstOrDefaultAsync(ct);
        return song is null ? ApiProblems.NotFound("Song", songId) : Results.Ok(WithLinks(song, request));
    }

    private static async Task<IResult> Create(int albumId, SongRequest body, GlazeAudioDbContext db, HttpRequest request, CancellationToken ct)
    {
        if (!await db.Albums.AnyAsync(a => a.Id == albumId, ct))
            return ApiProblems.NotFound("Album", albumId);

        var song = new Song
        {
            AlbumId = albumId,
            Title = body.Title!.Trim(),
            TrackNumber = body.TrackNumber!.Value,
            DurationSeconds = body.DurationSeconds!.Value
        };
        db.Songs.Add(song);
        await db.SaveChangesAsync(ct);

        var dto = new SongDto(song.Id, albumId, song.Title, song.TrackNumber, song.DurationSeconds, 0, null);
        return Results.CreatedAtRoute("GetSong", new { albumId, songId = song.Id }, WithLinks(dto, request));
    }

    private static async Task<IResult> Update(int albumId, int songId, SongRequest body, GlazeAudioDbContext db, HttpRequest request, CancellationToken ct)
    {
        if (!await db.Albums.AnyAsync(a => a.Id == albumId, ct))
            return ApiProblems.NotFound("Album", albumId);

        var song = await db.Songs.FirstOrDefaultAsync(s => s.AlbumId == albumId && s.Id == songId, ct);
        if (song is null) return ApiProblems.NotFound("Song", songId);

        song.Title = body.Title!.Trim();
        song.TrackNumber = body.TrackNumber!.Value;
        song.DurationSeconds = body.DurationSeconds!.Value;
        await db.SaveChangesAsync(ct);

        var dto = await Project(db.Songs.Where(s => s.Id == songId)).FirstAsync(ct);
        return Results.Ok(WithLinks(dto, request));
    }

    private static async Task<IResult> Delete(int albumId, int songId, GlazeAudioDbContext db, CancellationToken ct)
    {
        if (!await db.Albums.AnyAsync(a => a.Id == albumId, ct))
            return ApiProblems.NotFound("Album", albumId);

        var deleted = await db.Songs.Where(s => s.AlbumId == albumId && s.Id == songId).ExecuteDeleteAsync(ct);
        return deleted == 0 ? ApiProblems.NotFound("Song", songId) : Results.NoContent();
    }
}